[FIX] Move the login route definition ro app.py from run.py
